Bundle the results of: pip install Adafruit-Blinka==6.17.0 (armhf).
As installed on a Raspbian bullseye armhf image running on a Pi 4 B,
with the python3-rpi.gpio package purged, so that the list of files
(modulo architecture qualifiers) is exactly the same as the one for
arm64.

Ship everything under a private directory to make sure it doesn't
interfere with any existing Python modules.