Retire the pirogue-base-pc package.
We don't need to maintain two different metapackages on the long run: we
can let pirogue-base be the entry point for all machines, and then let
packages like pirogue-screen-st7789-240x240 and pirogue-hat deal with
hardware and software autodetection (see #12 and #14).

The exact details to help pirogue-base-pc upgrade from Debian 11 to
Debian 12 remain to be discussed (see #16).