Implement basic heuristics.
We need to distinguish between ARM systems, to make sure we tweak
DTB/DTBO settings based on the deployed distribution, and only if
hardware is actually a supported Raspberry Pi model (3 or 4).

On other systems, we can assume Debian is the deployed distribution,
as those don't need any hardware-specific configuration anyway.