Force all iptables-provided alternatives to auto (nft).
Given the number of reverse dependencies, it would be hard to declare
a conflict against iptables. Let's go with the flow, being happy if
it's not installed, but also trying to deal with it if it's installed.

In other words: we're neither pulling it via Depends nor making sure
it's not installed via Conflicts.

Interestingly, iptables itself ships:
 - legacy and nft implementation for iptables and ip6tables;
 - only the nft implementation for arptables and ebtables.

Meanwhile, the legacy implementation of arptables (resp. ebtables) is
shipped in the arptables (resp. ebtables) package.

This commit forces all 4 alternatives (iptables, ip6tables, arptables,
and ebtables) to auto mode. This should limit the risks of having
firewall rules deployed via the legacy interface while we focus on
nftables support (which has been the default in Debian for a few
releases already).

Do that once, when first deploying pirogue-networking.

Remarks:
 - Admins can always change the settings behind our back anyway (even
   if we're going to discourage toying with iptables via the PTS
   documentation), and we'll want to add some detection code to report
   iptables usage, which might conflict with our rules.
 - Some tools like ferm might use the legacy implementation anyway,
   without taking the alternatives system into account.

Link: https://github.com/PiRogueToolSuite/pirogue-admin/issues/15