Document last change in pirogue-maintenance.
Instead of using a systemd unit override that duplicates a value from
another package, this unit could leverage suricata.service's condition,
similarly to what's done in pirogue-eve-collector.postinst:

    if systemctl show suricata.service | grep -qs ^ConditionResult=yes$; then

But maybe that could be tricky if both units are evaluated in parallel
or in the wrong order. Since pirogue-maintenance.service runs via a
daily timer, that seems rather unlikely.